Update advanced.md
Hi! Thanks for this nice plugin, I was having some trouble activation the highlighting, because I missed this part of the `zsh-syntax-highlighting` docs. 

I think it might be useful to include something like this, or similar in the documentation to make people aware that you also have to do this.